scripts/requirements.txt
supabase==2.7.4
google-generativeai==0.7.2
httpx>=0.27.0
beautifulsoup4==4.12.3
pydantic==2.7.0
tenacity==8.2.3
rapidfuzz==3.8.1
rich==13.7.1
python-dotenv==1.0.1
apify-client==1.6.4
pytest==8.2.0
pytest-mock==3.14.0
